app/requirements.txt
Flask
Flask-Mail
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
alembic
boto3
gunicorn
